manual mode works, no timers
git-svn-id: https://openhr20.svn.sourceforge.net/svnroot/openhr20@16 e6282f77-c345-0410-9e5f-9d7f1dee61d8

+++ trunk/source/pid.h
@@ -40,4 +40,4 @@ void pid_Init(int16_t processValue);
  *  \param setPoint  Desired value.
  *  \param processValue  Measured value.
  */
-int16_t pid_Controller(int16_t setPoint, int16_t processValue);
+int8_t pid_Controller(int16_t setPoint, int16_t processValue);
